# docs/openapi.yaml
openapi: 3.0.3
info:
  title: spot-diggz API
  version: "0.1.0"
  description: |
    spot-diggz API contract (Rust/Axum). This is an internal contract for client development.
servers:
  - url: http://localhost:8080
    description: Local development
paths:
  /sdz/health:
    get:
      summary: Health check
      x-sdz-client:
        - web
        - ios
        - android
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
  /sdz/users/me:
    get:
      summary: Get current user profile
      security:
        - BearerAuth: []
      x-sdz-client:
        - web
        - ios
        - android
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/ErrorUnauthorized"
        "404":
          $ref: "#/components/responses/ErrorNotFound"
  /sdz/spots:
    post:
      summary: Create a skate spot (mobile only)
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ClientHeader"
      x-sdz-client:
        - ios
        - android
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSpotInput"
      responses:
        "200":
          description: Created spot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Spot"
        "400":
          $ref: "#/components/responses/ErrorBadRequest"
        "401":
          $ref: "#/components/responses/ErrorUnauthorized"
        "403":
          $ref: "#/components/responses/ErrorForbidden"
        "500":
          $ref: "#/components/responses/ErrorInternal"
    get:
      summary: List spots
      x-sdz-client:
        - web
        - ios
        - android
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Spot"
        "500":
          $ref: "#/components/responses/ErrorInternal"
  /sdz/spots/{spot_id}:
    get:
      summary: Get a spot by ID
      parameters:
        - name: spot_id
          in: path
          required: true
          schema:
            type: string
      x-sdz-client:
        - web
        - ios
        - android
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Spot"
        "404":
          $ref: "#/components/responses/ErrorNotFound"
        "500":
          $ref: "#/components/responses/ErrorInternal"
  /sdz/spots/upload-url:
    post:
      summary: Generate signed upload URL (mobile only)
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ClientHeader"
      x-sdz-client:
        - ios
        - android
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "400":
          $ref: "#/components/responses/ErrorBadRequest"
        "401":
          $ref: "#/components/responses/ErrorUnauthorized"
        "403":
          $ref: "#/components/responses/ErrorForbidden"
        "500":
          $ref: "#/components/responses/ErrorInternal"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ClientHeader:
      name: X-SDZ-Client
      in: header
      required: true
      description: Client type (`ios` or `android`) for mobile-only endpoints.
      schema:
        type: string
        enum: [ios, android]
  schemas:
    HealthResponse:
      type: object
      required: [status, message]
      properties:
        status:
          type: string
          example: healthy
        message:
          type: string
          example: spot-diggz api is running
    ErrorResponse:
      type: object
      required: [code, errorCode, message]
      properties:
        code:
          type: integer
          format: int32
          example: 401
        errorCode:
          type: string
          example: SDZ-E-1001
        message:
          type: string
          example: Unauthorized
    SpotLocation:
      type: object
      required: [lat, lng]
      properties:
        lat:
          type: number
          format: double
          example: 35.6812
        lng:
          type: number
          format: double
          example: 139.7671
    SpotTimeRange:
      type: object
      required: [startMinutes, endMinutes]
      properties:
        startMinutes:
          type: integer
          format: int32
          example: 540
        endMinutes:
          type: integer
          format: int32
          example: 1260
    SpotBusinessHours:
      type: object
      required: [is24Hours, sameAsWeekday]
      properties:
        scheduleType:
          type: string
          nullable: true
          enum:
            - regular
            - weekdayOnly
            - weekendOnly
            - irregular
            - schoolOnly
            - manual
        is24Hours:
          type: boolean
          example: false
        sameAsWeekday:
          type: boolean
          example: true
        weekday:
          $ref: "#/components/schemas/SpotTimeRange"
          nullable: true
        weekend:
          $ref: "#/components/schemas/SpotTimeRange"
          nullable: true
        note:
          type: string
          nullable: true
    SpotParkAttributes:
      type: object
      properties:
        officialUrl:
          type: string
          nullable: true
        businessHours:
          $ref: "#/components/schemas/SpotBusinessHours"
          nullable: true
        accessInfo:
          type: string
          nullable: true
        phoneNumber:
          type: string
          nullable: true
    SpotStreetSurfaceCondition:
      type: object
      properties:
        roughness:
          type: string
          nullable: true
        crack:
          type: string
          nullable: true
    SpotStreetSection:
      type: object
      required: [type]
      properties:
        type:
          type: string
        count:
          type: integer
          format: int32
          nullable: true
        heightCm:
          type: integer
          format: int32
          nullable: true
        widthCm:
          type: integer
          format: int32
          nullable: true
        notes:
          type: string
          nullable: true
    SpotStreetAttributes:
      type: object
      properties:
        surfaceMaterial:
          type: string
          nullable: true
        surfaceCondition:
          $ref: "#/components/schemas/SpotStreetSurfaceCondition"
          nullable: true
        sections:
          type: array
          items:
            $ref: "#/components/schemas/SpotStreetSection"
          nullable: true
        difficulty:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
    Spot:
      type: object
      required:
        - spotId
        - name
        - tags
        - images
        - userId
        - createdAt
        - updatedAt
      properties:
        spotId:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        location:
          $ref: "#/components/schemas/SpotLocation"
          nullable: true
        tags:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: string
          maxItems: 3
          description: 1スポット3枚まで。画像付きスポットは1ユーザーあたり最大10件まで。
        approvalStatus:
          type: string
          description: Spot approval status (pending/approved/rejected). Omitted when no request yet.
          enum:
            - pending
            - approved
            - rejected
        parkAttributes:
          $ref: "#/components/schemas/SpotParkAttributes"
          nullable: true
        streetAttributes:
          $ref: "#/components/schemas/SpotStreetAttributes"
          nullable: true
        instagramTag:
          type: string
          nullable: true
        userId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateSpotInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        location:
          $ref: "#/components/schemas/SpotLocation"
          nullable: true
        tags:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: string
          maxItems: 3
          description: 1スポット3枚まで
        parkAttributes:
          $ref: "#/components/schemas/SpotParkAttributes"
          nullable: true
        streetAttributes:
          $ref: "#/components/schemas/SpotStreetAttributes"
          nullable: true
        instagramTag:
          type: string
          nullable: true
    UploadUrlRequest:
      type: object
      required: [contentType]
      properties:
        contentType:
          type: string
          example: image/jpeg
    UploadUrlResponse:
      type: object
      required: [uploadUrl, objectUrl, objectName, expiresAt]
      properties:
        uploadUrl:
          type: string
        objectUrl:
          type: string
        objectName:
          type: string
        expiresAt:
          type: string
          format: date-time
    User:
      type: object
      required: [userId, displayName]
      properties:
        userId:
          type: string
        displayName:
          type: string
        email:
          type: string
          nullable: true
  responses:
    ErrorBadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ErrorUnauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ErrorForbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ErrorNotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ErrorInternal:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"